adjustments to the core do a make sure
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@3504 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/languages/mod_spidermonkey/mod_spidermonkey.c
@@ -862,14 +862,14 @@ static JSBool session_recordfile(JSContext *cx, JSObject *obj, uintN argc, jsval
         }
 
         if (argc > 4) {
-            int32_t thresh;
-            JS_ValueToInt32(cx, argv[4], &(int32)thresh);
+            int32 thresh;
+            JS_ValueToInt32(cx, argv[4], &thresh);
             fh.thresh = thresh;
         }
 
         if (argc > 5) {
-            int32_t silence_hits;
-            JS_ValueToInt32(cx, argv[5], &(int32)silence_hits);
+            int32 silence_hits;
+            JS_ValueToInt32(cx, argv[5], &silence_hits);
             fh.silence_hits = silence_hits;
         }
 	}
